+++ apps/doop/src/components/shared/HintLoading.tsx
@@ -9,7 +9,7 @@ import { Stack, Spinner } from "@cloudoperators/juno-ui-components"
 const HintLoading = ({ text, className }: any) => {
   return (
     <Stack alignment="center" className={className}>
-      <Spinner variant="primary" />
+      <Spinner variant="primary" className="mr-3" />
       {text ? <span>{text}</span> : <span>Loading...</span>}
     </Stack>
   )